requirements.txt
customtkinter
psutil
pyusb
cryptography
libusb
libusb-package
pywin32
